refactor: 도메인, 뷰 함수 분리
