+++ src/index.js
@@ -1,72 +1,10 @@
-import { generateRandomNumber, checkGuessNumber } from './utils/numbers.js';
-import { readLineAsync, getUserInput } from './utils/inputHandler.js';
-
-const createGameSetting = () => {
-  let minNumber = null;
-  let maxNumber = null;
-  let playNumber = null;
-
-  return {
-    askToMinMaxNumber: async () => {
-      console.log(
-        '[게임 설정] 게임 시작을 위해 최소 값, 최대 값을 입력해주세요. (예: 1, 50)'
-      );
-      const minAndMaxNumber = await readLineAsync('숫자 입력:');
-      if (!minAndMaxNumber.includes(',')) {
-        console.log('숫자 입력이 잘못되었습니다. 다시 입력해주세요.');
-        return askToMinMaxNumber();
-      }
-      if (minAndMaxNumber.split(',')[1] === '') {
-        console.log('숫자 입력이 잘못되었습니다. 다시 입력해주세요.');
-        return askToMinMaxNumber();
-      }
-      if (minAndMaxNumber.split(',').length !== 2) {
-        console.log('숫자 입력이 잘못되었습니다. 다시 입력해주세요.');
-        return askToMinMaxNumber();
-      }
-      [minNumber, maxNumber] = minAndMaxNumber.split(',').map(Number);
-      return [minNumber, maxNumber];
-    },
-    askToPlayNumber: async () => {
-      console.log(
-        '[게임 설정] 게임 시작을 위해 진행 가능 횟수를 입력해주세요.'
-      );
-      playNumber = await readLineAsync('숫자 입력:');
-      return Number(playNumber);
-    },
-
-    getSettings: () => ({
-      minNumber,
-      maxNumber,
-      playNumber,
-    }),
-  };
-};
-
-const createGameProcess = (randomNumber, playNumber) => {
-  let count = 0;
-  const inputValue = [];
-  return async function processGuess(userInput) {
-    inputValue.push(userInput);
-    count++;
-    checkGuessNumber(userInput, randomNumber, inputValue);
-    if (userInput === randomNumber) {
-      console.log(
-        `축하합니다! ${inputValue.length}번 만에 숫자를 맞추셨습니다.`
-      );
-      await askToPlayAgain();
-      return true;
-    }
-    if (count === playNumber) {
-      console.log(
-        `${count}회 초과! 숫자를 맞추지 못했습니다. (정답: ${randomNumber})`
-      );
-      await askToPlayAgain();
-      return true;
-    }
-    return false;
-  };
-};
+import {
+  createGameSetting,
+  createGameProcess,
+  generateRandomNumber,
+} from './domain/input.js';
+import { readLineAsync } from './utils/inputHandler.js';
+import { getUserInput } from './view/index.js';
 
 const play = async () => {
   const gameSettings = createGameSetting();
@@ -86,7 +24,7 @@ const play = async () => {
   }
 };
 
-const askToPlayAgain = async () => {
+export const askToPlayAgain = async () => {
   while (true) {
     const playAgain = await readLineAsync(
       '게임을 다시 시작하시겠습니까? (yes/no): '
